:recycle: Moved classes that can be shared outside the Shipment component
