+++ src/Shipment/Customs/Customs.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace MyParcelCom\Integration\Shipment\Customs;
 
-use MyParcelCom\Integration\Shipment\Price;
+use MyParcelCom\Integration\Price;
 use function array_filter;
 
 class Customs
