+++ src/Shipment/Items/Item.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace MyParcelCom\Integration\Shipment\Items;
 
-use MyParcelCom\Integration\Shipment\Price;
+use MyParcelCom\Integration\Price;
 use function array_filter;
 
 class Item
