+++ tests/Order/OrderTest.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Order;
+
+use PHPUnit\Framework\TestCase;
+
+class OrderTest extends TestCase
+{
+
+}
